refactor: ImageSelector Form에서 관리 및 폴더 이동
